+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - A new plugin **TDE2LevelStreamingUtils** was added to provide utilities (components/system) for basic level streaming based on bounding volumes.
 
+- A new functionality of loading user defined plugins that are defined in project settings was implemented.
+
 ## [0.6.4] 2022-05-25
 
 ### Added